+++ tests/consumers/add_ons/CMakeLists.txt
@@ -4,7 +4,7 @@ set(SOURCES
   ${dir}/FilterBlockTest.cpp
   ${dir}/periodic_trigger.cpp
   ${dir}/recent_rate.cpp
-  ${dir}/status.cpp
+  ${dir}/StatusTest.cpp
   ${dir}/ValueFilterTest.cpp
   ${dir}/ValueLatchTest.cpp
   )
